New and Improved
Bug Fix Log:

Program presses left when the controller is not plugged in.
- The controller was sending null when it was unplugged instead of a “0” value. The movement now checks if it is null as well and will not perform an action.

The program will not work over a remote connection. Controller keys will not work remotely with chrome remote desktop, It can not only run locally because of this. The region and cursor reference Images will not work remotely because there is too much compression to the original screen image, will not work locally because of this.
- We deployed half of the script remotely and half of the script locally to fix these issues.

The controller zooms in and out is too fast with the d-pad. Can we have different sensitivity for different axis?
- Created a custom zoom control for up and down compared to the right and left.

Need to update "READ ME" for remote desktop version install instructions.
-Updated for all installations