Updated Advanced Image Search Speed
Bug Fix Log (Edit 1-8)

Another way to check a bit faster would be to check the whole screen first and if we find a region, then go for the more precise region finder. This will make it faster when there are no regions on the screen, or command is run by accident. This will not solve the slowed speeds when they are far away.